Refactor the source layout for cuda.parallel
